Moved methods to proper activitys. Broken.
